implement basic language and interpreter
can now write a complete program in the text box and the backend will
run the program and spit out the status of the registers and the stack.
If there was an error, that too will be reported.
Currently limit the number of steps to 100,000 to stop infinite loops.
Otherwise, the page becomes unresponsive when the program doesnt terminate.